[DF-7335] Review fixes, more  nullptr instead of NULL in changed files.
Former-commit-id: 7ddbced44c1524e3f5964bceb0a6bc421008bc48

+++ Sources/Internal/Render/2D/TextLayout.cpp
@@ -152,7 +152,7 @@ bool TextLayout::NextBySymbols(const float32 lineWidth)
 {
     float32 targetWidth = std::floor(lineWidth);
     float32 currentLineDx = 0;
-    uint32 totalSize = static_cast<uint32>(preparedText.length());
+    size_t totalSize = preparedText.length();
     size_t pos = 0;
 
     for (pos = fromPos; pos < totalSize; pos++)
